+++ doc/changelog.md
@@ -3,6 +3,8 @@ jQuery-USOS Change Log
 
   * **BETA Branch** - *(unreleased)*
 
+    * **1.2.5.2** - Enhancements in the [usosTip widget](api/widget.tip.md),
+      e.g. the new `showAs` option.
     * **1.2.5.1** - **Requires USOS API 6.0.0.**
       New [badge widget](api/widget.badge.md): `entity/geo/building`.